Fix some types.

+++ modules/spaint/include/spaint/features/shared/VOPFeatureCalculator_Shared.h
@@ -49,8 +49,8 @@ _CPU_AND_GPU_CODE_
 inline void generate_coordinate_system(int voxelLocationIndex, const Vector3f *surfaceNormals, const unsigned int *voxelCountsForLabels,
                                        const size_t maxVoxelsPerLabel, Vector3f *xAxes, Vector3f *yAxes)
 {
-  unsigned int label = voxelLocationIndex / maxVoxelsPerLabel;
-  unsigned int offset = voxelLocationIndex % maxVoxelsPerLabel;
+  size_t label = voxelLocationIndex / maxVoxelsPerLabel;
+  size_t offset = voxelLocationIndex % maxVoxelsPerLabel;
   if(offset < voxelCountsForLabels[label])
   {
     Vector3f n = surfaceNormals[voxelLocationIndex];
@@ -68,8 +68,8 @@ inline void write_surface_normal(int voxelLocationIndex, const Vector3s *voxelLo
                                  const SpaintVoxel *voxelData, const ITMVoxelIndex::IndexData *indexData, const size_t maxVoxelsPerLabel,
                                  Vector3f *surfaceNormals)
 {
-  unsigned int label = voxelLocationIndex / maxVoxelsPerLabel;
-  unsigned int offset = voxelLocationIndex % maxVoxelsPerLabel;
+  size_t label = voxelLocationIndex / maxVoxelsPerLabel;
+  size_t offset = voxelLocationIndex % maxVoxelsPerLabel;
   if(offset < voxelCountsForLabels[label])
   {
     surfaceNormals[voxelLocationIndex] = computeSingleNormalFromSDF(voxelData, indexData, voxelLocations[voxelLocationIndex].toFloat());
